requirements.txt
PyPDF2==3.0.1
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.26.0
gradio==3.34.0
requests==2.31.0
pytesseract==0.3.10
Pillow==10.0.0
transformers==4.37.2

